+++ p2panda-rs/src/test_utils/fixtures/param_fixtures.rs
@@ -2,16 +2,19 @@
 
 /// General purpose fixtures which can be injected into rstest methods as parameters.
 ///
-/// The fixtures can optionally be passed in with custom parameters which overides the default values.
+/// The fixtures can optionally be passed in with custom parameters which overides the default
+/// values.
 use rstest::fixture;
 
 use crate::entry::{Entry, EntrySigned, SeqNum};
 use crate::hash::Hash;
 use crate::identity::KeyPair;
-use crate::operation::{Operation, OperationEncoded, OperationFields};
-use crate::test_utils::utils::{self, DEFAULT_HASH, DEFAULT_PRIVATE_KEY, DEFAULT_SCHEMA_HASH};
+use crate::operation::{Operation, OperationEncoded, OperationFields, OperationValue};
+use crate::test_utils::constants::{DEFAULT_HASH, DEFAULT_PRIVATE_KEY, DEFAULT_SCHEMA_HASH};
+use crate::test_utils::fixtures::defaults;
+use crate::test_utils::utils;
 
-/// Fixture struct which contains versioned p2panda data for testing
+/// Fixture struct which contains versioned p2panda data for testing.
 #[derive(Debug)]
 pub struct Fixture {
     pub entry: Entry,
@@ -20,14 +23,14 @@ pub struct Fixture {
     pub operation_encoded: OperationEncoded,
 }
 
-/// Fixture which injects the default private key string into a test method
+/// Fixture which injects the default private key string into a test method.
 #[fixture]
 pub fn private_key() -> String {
     DEFAULT_PRIVATE_KEY.into()
 }
 
-/// Fixture which injects the default KeyPair into a test method. Default value can be overridden at testing
-/// time by passing in a custom private key string.
+/// Fixture which injects the default KeyPair into a test method. Default value can be overridden
+/// at testing time by passing in a custom private key string.
 #[fixture]
 pub fn key_pair(private_key: String) -> KeyPair {
     utils::keypair_from_private(private_key)
@@ -39,38 +42,53 @@ pub fn random_key_pair() -> KeyPair {
     utils::new_key_pair()
 }
 
-/// Fixture which injects the default SeqNum into a test method. Default value can be overridden at testing
-/// time by passing in a custom seq num as i64.
+/// Fixture which injects the default SeqNum into a test method. Default value can be overridden at
+/// testing time by passing in a custom seq num as i64.
 #[fixture]
 pub fn seq_num(#[default(1)] n: i64) -> SeqNum {
     utils::seq_num(n)
 }
 
-/// Fixture which injects the default schema Hash into a test method. Default value can be overridden at testing
-/// time by passing in a custom schema hash string.
+/// Fixture which injects the default schema Hash into a test method. Default value can be
+/// overridden at testing time by passing in a custom schema hash string.
 #[fixture]
 pub fn schema(#[default(DEFAULT_SCHEMA_HASH)] schema_str: &str) -> Hash {
     utils::hash(schema_str)
 }
 
-/// Fixture which injects the default Hash into a test method. Default value can be overridden at testing
-/// time by passing in a custom hash string.
+/// Fixture which injects the default Hash into a test method. Default value can be overridden at
+/// testing time by passing in a custom hash string.
 #[fixture]
 pub fn hash(#[default(DEFAULT_HASH)] hash_str: &str) -> Hash {
     utils::hash(hash_str)
 }
 
-/// Fixture which injects the default OperationFields value into a test method. Default value can be overridden at testing
-/// time by passing in a custom vector of key-value tuples.
+/// Fixture which injects the default OperationFields value into a test method. Default value can
+/// be overridden at testing time by passing in a custom vector of key-value tuples.
 #[fixture]
 pub fn fields(
-    #[default(vec![("message", "Hello!")])] fields_vec: Vec<(&str, &str)>,
+    #[default(vec![("message", defaults::operation_value())])] fields_vec: Vec<(
+        &str,
+        OperationValue,
+    )>,
 ) -> OperationFields {
     utils::operation_fields(fields_vec)
 }
 
-/// Fixture which injects the default Entry into a test method. Default value can be overridden at testing
-/// time by passing in custom operation, seq number, backlink and skiplink.
+/// Fixture which injects the default OperationFields value into a test method. Default value can
+/// be overridden at testing time by passing in a custom vector of key-value tuples.
+#[fixture]
+pub fn some_fields(
+    #[default(vec![("message", defaults::operation_value())])] fields_vec: Vec<(
+        &str,
+        OperationValue,
+    )>,
+) -> Option<OperationFields> {
+    Some(utils::operation_fields(fields_vec))
+}
+
+/// Fixture which injects the default Entry into a test method. Default value can be overridden at
+/// testing time by passing in custom operation, seq number, backlink and skiplink.
 #[fixture]
 pub fn entry(
     operation: Operation,
@@ -81,55 +99,60 @@ pub fn entry(
     utils::entry(operation, skiplink, backlink, seq_num)
 }
 
-/// Fixture which injects the default Operation into a test method. Default value can be overridden at testing
-/// time by passing in custom operation fields and instance id.
+/// Fixture which injects the default Operation into a test method. Default value can be overridden
+/// at testing time by passing in custom operation fields and document id.
 #[fixture]
 pub fn operation(
-    #[default(Some(fields(vec![("message", "Hello!")])))] fields: Option<OperationFields>,
-    #[default(None)] instance_id: Option<Hash>,
+    #[from(some_fields)] fields: Option<OperationFields>,
+    #[default(None)] document_id: Option<Hash>,
 ) -> Operation {
-    utils::any_operation(fields, instance_id)
+    utils::any_operation(fields, document_id)
 }
 
-/// Fixture which injects the default Hash into a test method as an Option. Default value can be overridden at testing
-/// time by passing in custom hash string.
+/// Fixture which injects the default Hash into a test method as an Option. Default value can be
+/// overridden at testing time by passing in custom hash string.
 #[fixture]
 pub fn some_hash(#[default(DEFAULT_HASH)] str: &str) -> Option<Hash> {
     let hash = Hash::new(str);
     Some(hash.unwrap())
 }
 
-/// Fixture which injects the default CREATE Operation into a test method. Default value can be overridden at testing
-/// time by passing in custom schema hash and operation fields.
+/// Fixture which injects the default CREATE Operation into a test method. Default value can be
+/// overridden at testing time by passing in custom schema hash and operation fields.
 #[fixture]
 pub fn create_operation(schema: Hash, fields: OperationFields) -> Operation {
     utils::create_operation(schema, fields)
 }
 
-/// Fixture which injects the default UPDATE Operation into a test method. Default value can be overridden at testing
-/// time by passing in custom schema hash, instance id hash and operation fields.
+/// Fixture which injects the default UPDATE Operation into a test method. Default value can be
+/// overridden at testing time by passing in custom schema hash, document id hash and operation
+/// fields.
 #[fixture]
 pub fn update_operation(
     schema: Hash,
-    #[from(hash)] instance_id: Hash,
-    #[default(fields(vec![("message", "Updated, hello!")]))] fields: OperationFields,
+    #[from(hash)] document_id: Hash,
+    #[default(fields(vec![("message", OperationValue::Text("Updated, hello!".to_string()))]))]
+    fields: OperationFields,
 ) -> Operation {
-    utils::update_operation(schema, instance_id, fields)
+    utils::update_operation(schema, document_id, fields)
 }
 
-/// Fixture which injects the default DELETE Operation into a test method. Default value can be overridden at testing
-/// time by passing in custom schema hash and instance id hash.
+/// Fixture which injects the default DELETE Operation into a test method. Default value can be
+/// overridden at testing time by passing in custom schema hash and document id hash.
 #[fixture]
-pub fn delete_operation(schema: Hash, #[from(hash)] instance_id: Hash) -> Operation {
-    utils::delete_operation(schema, instance_id)
+pub fn delete_operation(schema: Hash, #[from(hash)] document_id: Hash) -> Operation {
+    utils::delete_operation(schema, document_id)
 }
 
-/// Fixture which injects p2panda testing data from p2panda version 0.3.0
+/// Fixture which injects p2panda testing data from p2panda version 0.3.0.
 #[fixture]
 pub fn v0_3_0_fixture() -> Fixture {
     let operation_fields = utils::operation_fields(vec![
-        ("name", "chess"),
-        ("description", "for playing chess"),
+        ("name", OperationValue::Text("chess".to_string())),
+        (
+            "description",
+            OperationValue::Text("for playing chess".to_string()),
+        ),
     ]);
     let operation = create_operation(Hash::new(DEFAULT_SCHEMA_HASH).unwrap(), operation_fields);
     let key_pair = utils::keypair_from_private(